Move the custom element definition to a separate file
